+++ src/model/auth/connect_account.ts
@@ -1,0 +1,11 @@
+export interface ConnectAccount {
+  account_type: ConnectType;
+  name: string;
+  email: string;
+}
+
+export enum ConnectType {
+  Google,
+  Facebook,
+  // TaiwanCloudEducation,
+}
