+++ src/model/auth/user.ts
@@ -1,4 +1,6 @@
 import { Schema } from 'mongoose';
+import { UserMode } from '@/model/auth/user_mode';
+import { ConnectAccount } from '@/model/auth/connect_account';
 
 export interface IUser {
   username: string;
@@ -10,33 +12,15 @@ export interface IUser {
   login_ips: string[];
 }
 
-export interface ConnectAccount {
-  account_type: ConnectType;
-  name: string;
-  email: string;
-}
-
-export enum ConnectType {
-  Google,
-  Facebook,
-  // TaiwanCloudEducation,
-}
-
-export enum UserMode {
-  Student,
-  Teacher,
-  Parents,
-}
-
 export const userSchema = new Schema<IUser>({
   username: { type: String, required: true },
   email: { type: String, required: true },
   verified_email: { type: Boolean, required: true },
   password_hash: { type: String, required: false },
-    connects: {
-        type: [{
-      
-  }], required: true },
+  connects: {
+    type: [{}],
+    required: true,
+  },
   modes: { type: [String], required: true },
   login_ips: { type: [String], required: true },
 });
