+++ src/model/auth/user_mode.ts
@@ -1,0 +1,5 @@
+export enum UserMode {
+  Student,
+  Teacher,
+  Parents,
+}
